# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Z-Fantasy Telegram Mini App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/create-invoice:
    post:
      operationId: createInvoice
      tags: [payments]
      summary: Create a Telegram Stars invoice link for a subscription tier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceRequest"
      responses:
        "200":
          description: Invoice link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
  /payments/webhook:
    post:
      operationId: handlePaymentWebhook
      tags: [payments]
      summary: Handle Telegram Stars payment success webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentWebhookPayload"
      responses:
        "200":
          description: Acknowledged
...
components:
  schemas:
...
    InvoiceRequest:
      type: object
      required: [tier, period]
      properties:
        tier: { type: string, enum: [Bronze, Silver, Gold] }
        period: { type: string, enum: [weekly, monthly, yearly] }
    InvoiceResponse:
      type: object
      required: [invoiceLink]
      properties:
        invoiceLink: { type: string }
    PaymentWebhookPayload:
      type: object
      properties:
        update_id: { type: integer }
        pre_checkout_query: { type: object }
        message: { type: object }
...