# lib/api-zod/src/generated/types/invoiceResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Z-Fantasy Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceResponse {
  invoiceLink: string;
}